extract Currency and tax choice types for reusability
