+++ src/PrestaShopBundle/Resources/config/services/core/form/choice_provider.yml
@@ -20,14 +20,6 @@ services:
       - '@=service("prestashop.adapter.legacy.context").getLanguage().id'
       - '@prestashop.adapter.data_provider.country'
 
-  PrestaShop\PrestaShop\Core\Form\ChoiceProvider\CurrencyByIdChoiceProvider:
-    autowire: true
-    public: false
-
-  prestashop.core.form.choice_provider.currency_by_id:
-    alias: PrestaShop\PrestaShop\Core\Form\ChoiceProvider\CurrencyByIdChoiceProvider
-    deprecated: ~
-
   prestashop.core.form.choice_provider.timezone_by_name:
     class: 'PrestaShop\PrestaShop\Core\Form\ChoiceProvider\TimezoneByNameChoiceProvider'
     arguments:
@@ -225,11 +217,6 @@ services:
     arguments:
       - '@translator'
 
-  prestashop.core.form.choice_provider.tax_inclusion:
-    class: 'PrestaShop\PrestaShop\Core\Form\ChoiceProvider\TaxInclusionChoiceProvider'
-    arguments:
-      - '@translator'
-
   prestashop.core.form.choice_provider.address_required_fields:
     class: 'PrestaShop\PrestaShop\Core\Form\ChoiceProvider\AddressRequiredFieldsChoiceProvider'
 
@@ -333,3 +320,22 @@ services:
   PrestaShop\PrestaShop\Core\Form\ChoiceProvider\DiscountApplicationChoiceProvider:
     autowire: true
     public: false
+
+  PrestaShop\PrestaShop\Core\Form\ChoiceProvider\TaxInclusionChoiceProvider:
+    autowire: true
+    public: false
+
+  PrestaShop\PrestaShop\Core\Form\ChoiceProvider\CurrencyByIdChoiceProvider:
+    autowire: true
+    public: false
+
+  # deprecated services
+  prestashop.core.form.choice_provider.currency_by_id:
+    alias: PrestaShop\PrestaShop\Core\Form\ChoiceProvider\CurrencyByIdChoiceProvider
+    public: true
+    deprecated: ~
+
+  prestashop.core.form.choice_provider.tax_inclusion:
+    alias: PrestaShop\PrestaShop\Core\Form\ChoiceProvider\TaxInclusionChoiceProvider
+    public: true
+    deprecated: ~
